% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{meditations_quotes}
\alias{meditations_quotes}
\title{All 507 quotes from Marcus Aurelius' book Meditations}
\format{A data frame with 53940 rows and 10 variables:
\describe{
  \item{quotes}{The text of the quote}
  ...
}}
\source{
\url{http://classics.mit.edu//Antoninus/meditations.1.one.html}
}
\usage{
meditations_quotes
}
\description{
A dataset containing the complete list of quotes of Marcus Aurelius from Meditations
}
\keyword{datasets}
