% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{meditations}
\alias{meditations}
\title{Get a random quote for Marcus Aurelius' book Meditations}
\usage{
meditations()
}
\value{
A string containing the quote.
}
\description{
Get a random quote for Marcus Aurelius' book Meditations
}
\examples{
meditations()
}
